move foreign call logic to separate module
