+++ acvm-repo/brillig_vm/src/lib.rs
@@ -12,8 +12,8 @@ use std::collections::HashMap;
 
 use acir::AcirField;
 use acir::brillig::{
-    BinaryFieldOp, BinaryIntOp, BitSize, ForeignCallParam, ForeignCallResult, HeapArray,
-    HeapValueType, HeapVector, IntegerBitSize, MemoryAddress, Opcode, ValueOrArray,
+    BinaryFieldOp, BinaryIntOp, ForeignCallParam, ForeignCallResult, IntegerBitSize, MemoryAddress,
+    Opcode,
 };
 use acvm_blackbox_solver::BlackBoxFunctionSolver;
 use arithmetic::{BrilligArithmeticError, evaluate_binary_field_op, evaluate_binary_int_op};
@@ -28,6 +28,7 @@ use num_bigint::BigUint;
 mod arithmetic;
 mod black_box;
 mod cast;
+mod foreign_call;
 mod memory;
 
 /// The error call stack contains the opcode indexes of the call stack at the time of failure, plus the index of the opcode that failed.
@@ -220,16 +221,6 @@ impl<'a, F: AcirField, B: BlackBoxFunctionSolver<F>> VM<'a, F, B> {
         self.status(VMStatus::Finished { return_data_offset, return_data_size })
     }
 
-    /// Sets the status of the VM to `ForeignCallWait`.
-    /// Indicating that the VM is now waiting for a foreign call to be resolved.
-    fn wait_for_foreign_call(
-        &mut self,
-        function: String,
-        inputs: Vec<ForeignCallParam<F>>,
-    ) -> VMStatus<F> {
-        self.status(VMStatus::ForeignCallWait { function, inputs })
-    }
-
     /// Provide the results of a Foreign Call to the VM
     /// and resume execution of the VM.
     pub fn resolve_foreign_call(&mut self, foreign_call_result: ForeignCallResult<F>) {
@@ -276,6 +267,14 @@ impl<'a, F: AcirField, B: BlackBoxFunctionSolver<F>> VM<'a, F, B> {
 
     pub fn get_memory(&self) -> &[MemoryValue<F>] {
         self.memory.values()
+    }
+
+    pub fn take_memory(mut self) -> Memory<F> {
+        std::mem::take(&mut self.memory)
+    }
+
+    pub fn foreign_call_counter(&self) -> usize {
+        self.foreign_call_counter
     }
 
     pub fn write_memory_at(&mut self, ptr: usize, value: MemoryValue<F>) {
@@ -499,78 +498,13 @@ impl<'a, F: AcirField, B: BlackBoxFunctionSolver<F>> VM<'a, F, B> {
                 destination_value_types,
                 inputs,
                 input_value_types,
-            } => {
-                assert!(inputs.len() == input_value_types.len());
-                assert!(destinations.len() == destination_value_types.len());
-
-                if self.foreign_call_counter >= self.foreign_call_results.len() {
-                    // When this opcode is called, it is possible that the results of a foreign call are
-                    // not yet known (not enough entries in `foreign_call_results`).
-                    // If that is the case, just resolve the inputs and pause the VM with a status
-                    // (VMStatus::ForeignCallWait) that communicates the foreign function name and
-                    // resolved inputs back to the caller. Once the caller pushes to `foreign_call_results`,
-                    // they can then make another call to the VM that starts at this opcode
-                    // but has the necessary results to proceed with execution.
-
-                    // With slices we might have more items in the HeapVector than the semantic length
-                    // indicated by the field preceding the pointer to the vector in the inputs.
-                    // This happens when SSA merges slices of different length, which can result in
-                    // a vector that has room for the longer of the two, partially filled with items
-                    // from the shorter. There are ways to deal with this on the receiver side,
-                    // but it is cumbersome, and the cleanest solution is not to send the extra empty
-                    // items at all. To do this, however, we need infer which input is the vector length.
-                    let mut vector_length: Option<usize> = None;
-
-                    let resolved_inputs = inputs
-                        .iter()
-                        .zip(input_value_types)
-                        .map(|(input, input_type)| {
-                            let mut input = self.get_memory_values(*input, input_type);
-                            // Truncate slices to their semantic length, which we remember from the preceding field.
-                            match input_type {
-                                HeapValueType::Simple(BitSize::Integer(IntegerBitSize::U32)) => {
-                                    // If we have a single u32 we may have a slice representation, so store this input.
-                                    // On the next iteration, if we have a vector then we know we have the dynamic length
-                                    // for that slice.
-                                    let ForeignCallParam::Single(length) = input else {
-                                        unreachable!("expected u32; got {input:?}");
-                                    };
-                                    vector_length = Some(length.to_u128() as usize);
-                                }
-                                HeapValueType::Vector { value_types } => {
-                                    if let Some(length) = vector_length {
-                                        let type_size = vector_element_size(value_types);
-                                        let mut fields = input.fields();
-                                        fields.truncate(length * type_size);
-                                        input = ForeignCallParam::Array(fields);
-                                    }
-                                    vector_length = None;
-                                }
-                                _ => {
-                                    // Otherwise we are not dealing with a u32 followed by a vector.
-                                    vector_length = None;
-                                }
-                            }
-                            input
-                        })
-                        .collect::<Vec<_>>();
-
-                    return self.wait_for_foreign_call(function.clone(), resolved_inputs);
-                }
-
-                let write_result = self.write_foreign_call_result(
-                    destinations,
-                    destination_value_types,
-                    self.foreign_call_counter,
-                );
-
-                if let Err(e) = write_result {
-                    return self.fail(e);
-                }
-
-                self.foreign_call_counter += 1;
-                self.increment_program_counter()
-            }
+            } => self.process_foreign_call(
+                function,
+                destinations,
+                destination_value_types,
+                inputs,
+                input_value_types,
+            ),
             Opcode::Mov { destination: destination_address, source: source_address } => {
                 let source_value = self.memory.read(*source_address);
                 self.memory.write(*destination_address, source_value);
@@ -675,411 +609,6 @@ impl<'a, F: AcirField, B: BlackBoxFunctionSolver<F>> VM<'a, F, B> {
         self.status.clone()
     }
 
-    /// Get input data from memory to pass to foreign calls.
-    fn get_memory_values(
-        &self,
-        input: ValueOrArray,
-        value_type: &HeapValueType,
-    ) -> ForeignCallParam<F> {
-        match (input, value_type) {
-            (ValueOrArray::MemoryAddress(value_index), HeapValueType::Simple(_)) => {
-                ForeignCallParam::Single(self.memory.read(value_index).to_field())
-            }
-            (
-                ValueOrArray::HeapArray(HeapArray { pointer: pointer_index, size }),
-                HeapValueType::Array { value_types, size: type_size },
-            ) if *type_size == size => {
-                let start = self.memory.read_ref(pointer_index);
-                self.read_slice_of_values_from_memory(start, size, value_types)
-                    .into_iter()
-                    .map(|mem_value| mem_value.to_field())
-                    .collect::<Vec<_>>()
-                    .into()
-            }
-            (
-                ValueOrArray::HeapVector(HeapVector { pointer: pointer_index, size: size_index }),
-                HeapValueType::Vector { value_types },
-            ) => {
-                let start = self.memory.read_ref(pointer_index);
-                let size = self.memory.read(size_index).to_usize();
-                self.read_slice_of_values_from_memory(start, size, value_types)
-                    .into_iter()
-                    .map(|mem_value| mem_value.to_field())
-                    .collect::<Vec<_>>()
-                    .into()
-            }
-            _ => {
-                unreachable!("Unexpected value type {value_type:?} for input {input:?}");
-            }
-        }
-    }
-
-    /// Reads an array/vector from memory but recursively reads pointers to
-    /// nested arrays/vectors according to the sequence of value types.
-    fn read_slice_of_values_from_memory(
-        &self,
-        start: MemoryAddress,
-        size: usize,
-        value_types: &[HeapValueType],
-    ) -> Vec<MemoryValue<F>> {
-        assert!(!start.is_relative(), "read_slice_of_values_from_memory requires direct addresses");
-        if HeapValueType::all_simple(value_types) {
-            self.memory.read_slice(start, size).to_vec()
-        } else {
-            // Check that the sequence of value types fit an integer number of
-            // times inside the given size.
-            assert!(
-                0 == size % value_types.len(),
-                "array/vector does not contain a whole number of elements"
-            );
-
-            // We want to send vectors to foreign functions truncated to their semantic length.
-            let mut vector_length: Option<usize> = None;
-
-            (0..size)
-                .zip(value_types.iter().cycle())
-                .map(|(i, value_type)| {
-                    let value_address = start.offset(i);
-                    let values = match value_type {
-                        HeapValueType::Simple(_) => {
-                            vec![self.memory.read(value_address)]
-                        }
-                        HeapValueType::Array { value_types, size } => {
-                            let array_address = self.memory.read_ref(value_address);
-
-                            self.read_slice_of_values_from_memory(
-                                array_address.offset(1),
-                                *size,
-                                value_types,
-                            )
-                        }
-                        HeapValueType::Vector { value_types } => {
-                            let vector_address = self.memory.read_ref(value_address);
-                            let size_address =
-                                MemoryAddress::direct(vector_address.unwrap_direct() + 1);
-                            let items_start = vector_address.offset(2);
-                            let vector_size = self.memory.read(size_address).to_usize();
-                            self.read_slice_of_values_from_memory(
-                                items_start,
-                                vector_size,
-                                value_types,
-                            )
-                        }
-                    };
-                    (value_type, values)
-                })
-                .flat_map(|(value_type, mut values)| {
-                    match value_type {
-                        HeapValueType::Simple(BitSize::Integer(IntegerBitSize::U32)) => {
-                            vector_length = Some(values[0].to_usize());
-                        }
-                        HeapValueType::Vector { value_types } => {
-                            if let Some(length) = vector_length {
-                                let type_size = vector_element_size(value_types);
-                                values.truncate(length * type_size);
-                            }
-                            vector_length = None;
-                        }
-                        _ => {
-                            // Otherwise we are not dealing with a u32 followed by a vector.
-                            vector_length = None;
-                        }
-                    }
-                    values
-                })
-                .collect::<Vec<_>>()
-        }
-    }
-
-    /// Write a foreign call's results to the VM memory.
-    ///
-    /// We match the expected types with the actual results.
-    /// However foreign call results do not support nested structures:
-    /// They are either a single integer value or a vector of integer values (field elements).
-    /// Therefore, nested arrays returned from foreign call results are flattened.
-    /// If the expected array sizes do not match the actual size, we reconstruct the nested
-    /// structure from the flat output array.
-    fn write_foreign_call_result(
-        &mut self,
-        destinations: &[ValueOrArray],
-        destination_value_types: &[HeapValueType],
-        foreign_call_index: usize,
-    ) -> Result<(), String> {
-        let values = std::mem::take(&mut self.foreign_call_results[foreign_call_index].values);
-
-        if destinations.len() != values.len() {
-            return Err(format!(
-                "{} output values were provided as a foreign call result for {} destination slots",
-                values.len(),
-                destinations.len()
-            ));
-        }
-
-        debug_assert_eq!(
-            destinations.len(),
-            destination_value_types.len(),
-            "Number of destinations must match number of value types",
-        );
-        debug_assert_eq!(
-            destinations.len(),
-            values.len(),
-            "Number of foreign call return values must match number of destinations",
-        );
-        for ((destination, value_type), output) in
-            destinations.iter().zip(destination_value_types).zip(&values)
-        {
-            match (destination, value_type) {
-                (ValueOrArray::MemoryAddress(value_index), HeapValueType::Simple(bit_size)) => {
-                    let output_fields = output.fields();
-                    if value_type
-                        .flattened_size()
-                        .is_some_and(|flattened_size| output_fields.len() != flattened_size)
-                    {
-                        return Err(format!(
-                            "Foreign call return value does not match expected size. Expected {} but got {}",
-                            value_type.flattened_size().unwrap(),
-                            output_fields.len(),
-                        ));
-                    }
-
-                    match output {
-                        ForeignCallParam::Single(value) => {
-                            self.write_value_to_memory(*value_index, value, *bit_size)?;
-                        }
-                        _ => {
-                            return Err(format!(
-                                "Function result size does not match brillig bytecode. Expected 1 result but got {output:?}"
-                            ));
-                        }
-                    }
-                }
-                (
-                    ValueOrArray::HeapArray(HeapArray { pointer: pointer_index, size }),
-                    HeapValueType::Array { value_types, size: type_size },
-                ) if size == type_size => {
-                    let output_fields = output.fields();
-                    if value_type
-                        .flattened_size()
-                        .is_some_and(|flattened_size| output_fields.len() != flattened_size)
-                    {
-                        return Err(format!(
-                            "Foreign call return value does not match expected size. Expected {} but got {}",
-                            value_type.flattened_size().unwrap(),
-                            output_fields.len(),
-                        ));
-                    }
-
-                    if HeapValueType::all_simple(value_types) {
-                        match output {
-                            ForeignCallParam::Array(values) => {
-                                if values.len() != *size {
-                                    // foreign call returning flattened values into a nested type, so the sizes do not match
-                                    let destination = self.memory.read_ref(*pointer_index);
-
-                                    let mut flatten_values_idx = 0; //index of values read from flatten_values
-                                    self.write_slice_of_values_to_memory(
-                                        destination,
-                                        &output_fields,
-                                        &mut flatten_values_idx,
-                                        value_type,
-                                    )?;
-                                    // Should be caught earlier but we want to be explicit.
-                                    debug_assert_eq!(
-                                        flatten_values_idx,
-                                        output_fields.len(),
-                                        "Not all values were written to memory"
-                                    );
-                                } else {
-                                    self.write_values_to_memory_slice(
-                                        *pointer_index,
-                                        values,
-                                        value_types,
-                                    )?;
-                                }
-                            }
-                            _ => {
-                                return Err(
-                                    "Function result size does not match brillig bytecode size"
-                                        .to_string(),
-                                );
-                            }
-                        }
-                    } else {
-                        // foreign call returning flattened values into a nested type, so the sizes do not match
-                        let destination = self.memory.read_ref(*pointer_index);
-                        let return_type = value_type;
-                        let mut flatten_values_idx = 0; //index of values read from flatten_values
-                        self.write_slice_of_values_to_memory(
-                            destination,
-                            &output_fields,
-                            &mut flatten_values_idx,
-                            return_type,
-                        )?;
-                        debug_assert_eq!(
-                            flatten_values_idx,
-                            output_fields.len(),
-                            "Not all values were written to memory"
-                        );
-                    }
-                }
-                (
-                    ValueOrArray::HeapVector(HeapVector {
-                        pointer: pointer_index,
-                        size: size_index,
-                    }),
-                    HeapValueType::Vector { value_types },
-                ) => {
-                    if HeapValueType::all_simple(value_types) {
-                        match output {
-                            ForeignCallParam::Array(values) => {
-                                if values.len() % value_types.len() != 0 {
-                                    return Err("Returned data does not match vector element size"
-                                        .to_string());
-                                }
-                                // Set our size in the size address
-                                self.memory.write(*size_index, values.len().into());
-                                self.write_values_to_memory_slice(
-                                    *pointer_index,
-                                    values,
-                                    value_types,
-                                )?;
-                            }
-                            _ => {
-                                return Err(
-                                    "Function result size does not match brillig bytecode size"
-                                        .to_string(),
-                                );
-                            }
-                        }
-                    } else {
-                        unimplemented!("deflattening heap vectors from foreign calls");
-                    }
-                }
-                _ => {
-                    return Err(format!(
-                        "Unexpected value type {value_type:?} for destination {destination:?}"
-                    ));
-                }
-            }
-        }
-
-        self.foreign_call_results[foreign_call_index].values = values;
-
-        Ok(())
-    }
-
-    fn write_value_to_memory(
-        &mut self,
-        destination: MemoryAddress,
-        value: &F,
-        value_bit_size: BitSize,
-    ) -> Result<(), String> {
-        let memory_value = MemoryValue::new_checked(*value, value_bit_size);
-
-        if let Some(memory_value) = memory_value {
-            self.memory.write(destination, memory_value);
-        } else {
-            return Err(format!(
-                "Foreign call result value {value} does not fit in bit size {value_bit_size:?}"
-            ));
-        }
-        Ok(())
-    }
-
-    fn write_values_to_memory_slice(
-        &mut self,
-        pointer_index: MemoryAddress,
-        values: &[F],
-        value_types: &[HeapValueType],
-    ) -> Result<(), String> {
-        let bit_sizes_iterator = value_types
-            .iter()
-            .map(|typ| match typ {
-                HeapValueType::Simple(bit_size) => *bit_size,
-                _ => unreachable!("Expected simple value type"),
-            })
-            .cycle();
-
-        // Convert the destination pointer to a usize
-        let destination = self.memory.read_ref(pointer_index);
-        // Write to our destination memory
-        let memory_values: Option<Vec<_>> = values
-            .iter()
-            .zip(bit_sizes_iterator)
-            .map(|(value, bit_size)| MemoryValue::new_checked(*value, bit_size))
-            .collect();
-        if let Some(memory_values) = memory_values {
-            self.memory.write_slice(destination, &memory_values);
-        } else {
-            return Err(format!(
-                "Foreign call result values {values:?} do not match expected bit sizes",
-            ));
-        }
-        Ok(())
-    }
-
-    /// Writes flattened values to memory, using the provided type
-    /// Function calls itself recursively in order to work with recursive types (nested arrays)
-    /// values_idx is the current index in the values vector and is incremented every time
-    /// a value is written to memory
-    /// The function returns the address of the next value to be written
-    fn write_slice_of_values_to_memory(
-        &mut self,
-        destination: MemoryAddress,
-        values: &Vec<F>,
-        values_idx: &mut usize,
-        value_type: &HeapValueType,
-    ) -> Result<(), String> {
-        assert!(
-            !destination.is_relative(),
-            "write_slice_of_values_to_memory requires direct addresses"
-        );
-        let mut current_pointer = destination;
-        match value_type {
-            HeapValueType::Simple(bit_size) => {
-                self.write_value_to_memory(destination, &values[*values_idx], *bit_size)?;
-                *values_idx += 1;
-                Ok(())
-            }
-            HeapValueType::Array { value_types, size } => {
-                for _ in 0..*size {
-                    for typ in value_types {
-                        match typ {
-                            HeapValueType::Simple(len) => {
-                                self.write_value_to_memory(
-                                    current_pointer,
-                                    &values[*values_idx],
-                                    *len,
-                                )?;
-                                *values_idx += 1;
-                                current_pointer = current_pointer.offset(1);
-                            }
-                            HeapValueType::Array { .. } => {
-                                let destination = self.memory.read_ref(current_pointer).offset(1);
-                                self.write_slice_of_values_to_memory(
-                                    destination,
-                                    values,
-                                    values_idx,
-                                    typ,
-                                )?;
-                                current_pointer = current_pointer.offset(1);
-                            }
-                            HeapValueType::Vector { .. } => {
-                                return Err(format!(
-                                    "Unsupported returned type in foreign calls {typ:?}"
-                                ));
-                            }
-                        }
-                    }
-                }
-                Ok(())
-            }
-            HeapValueType::Vector { .. } => {
-                Err(format!("Unsupported returned type in foreign calls {value_type:?}"))
-            }
-        }
-    }
-
     /// Process a binary operation.
     /// This method will not modify the program counter.
     fn process_binary_field_op(
@@ -1140,23 +669,13 @@ impl<'a, F: AcirField, B: BlackBoxFunctionSolver<F>> VM<'a, F, B> {
     }
 }
 
-/// Returns the total number of field elements required to represent the elements in the vector in memory.
-///
-/// Panics if the vector contains nested vectors. Such types are not supported and are rejected by the frontend.
-fn vector_element_size(value_types: &[HeapValueType]) -> usize {
-    value_types
-        .iter()
-        .map(|typ| {
-            typ.flattened_size()
-                .unwrap_or_else(|| panic!("unexpected nested dynamic element type: {typ:?}"))
-        })
-        .sum()
-}
-
 #[cfg(test)]
 mod tests {
     use crate::memory::MEMORY_ADDRESSING_BIT_SIZE;
-    use acir::{AcirField, FieldElement};
+    use acir::{
+        AcirField, FieldElement,
+        brillig::{BitSize, HeapVector},
+    };
     use acvm_blackbox_solver::StubbedBlackBoxSolver;
 
     use super::*;
@@ -1923,646 +1442,6 @@ mod tests {
             }
             assert_eq!(status, VMStatus::InProgress);
         }
-    }
-
-    #[test]
-    fn foreign_call_opcode_simple_result() {
-        let r_input = MemoryAddress::direct(0);
-        let r_result = MemoryAddress::direct(1);
-
-        let double_program = vec![
-            // Load input address with value 5
-            Opcode::Const {
-                destination: r_input,
-                value: (5u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // Call foreign function "double" with the input address
-            Opcode::ForeignCall {
-                function: "double".into(),
-                destinations: vec![ValueOrArray::MemoryAddress(r_result)],
-                destination_value_types: vec![HeapValueType::Simple(BitSize::Integer(
-                    MEMORY_ADDRESSING_BIT_SIZE,
-                ))],
-                inputs: vec![ValueOrArray::MemoryAddress(r_input)],
-                input_value_types: vec![HeapValueType::Simple(BitSize::Integer(
-                    MEMORY_ADDRESSING_BIT_SIZE,
-                ))],
-            },
-        ];
-
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm = brillig_execute_and_get_vm(vec![], &double_program, &solver);
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "double".into(),
-                inputs: vec![FieldElement::from(5usize).into()]
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(
-            FieldElement::from(10u128).into(), // Result of doubling 5u128
-        );
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check result address
-        let result_value = vm.memory.read(r_result);
-        assert_eq!(result_value, (10u32).into());
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
-    }
-
-    #[test]
-    fn foreign_call_opcode_memory_result() {
-        let r_input = MemoryAddress::direct(0);
-        let r_output = MemoryAddress::direct(1);
-
-        // Define a simple 2x2 matrix in memory
-        let initial_matrix: Vec<FieldElement> =
-            vec![(1u128).into(), (2u128).into(), (3u128).into(), (4u128).into()];
-
-        // Transpose of the matrix (but arbitrary for this test, the 'correct value')
-        let expected_result: Vec<FieldElement> =
-            vec![(1u128).into(), (3u128).into(), (2u128).into(), (4u128).into()];
-
-        let invert_program = vec![
-            Opcode::Const {
-                destination: MemoryAddress::direct(0),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(initial_matrix.len() as u32),
-            },
-            Opcode::Const {
-                destination: MemoryAddress::direct(1),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(0u64),
-            },
-            Opcode::CalldataCopy {
-                destination_address: MemoryAddress::direct(2),
-                size_address: MemoryAddress::direct(0),
-                offset_address: MemoryAddress::direct(1),
-            },
-            // input = 0
-            Opcode::Const {
-                destination: r_input,
-                value: 2_usize.into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output = 0
-            Opcode::Const {
-                destination: r_output,
-                value: 2_usize.into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // *output = matrix_2x2_transpose(*input)
-            Opcode::ForeignCall {
-                function: "matrix_2x2_transpose".into(),
-                destinations: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_output,
-                    size: initial_matrix.len(),
-                })],
-                destination_value_types: vec![HeapValueType::Array {
-                    size: initial_matrix.len(),
-                    value_types: vec![HeapValueType::field()],
-                }],
-                inputs: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_input,
-                    size: initial_matrix.len(),
-                })],
-                input_value_types: vec![HeapValueType::Array {
-                    value_types: vec![HeapValueType::field()],
-                    size: initial_matrix.len(),
-                }],
-            },
-        ];
-
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm = brillig_execute_and_get_vm(initial_matrix.clone(), &invert_program, &solver);
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "matrix_2x2_transpose".into(),
-                inputs: vec![initial_matrix.into()]
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(expected_result.clone().into());
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check result in memory
-        let result_values = vm.memory.read_slice(MemoryAddress::direct(2), 4).to_vec();
-        assert_eq!(
-            result_values.into_iter().map(|mem_value| mem_value.to_field()).collect::<Vec<_>>(),
-            expected_result
-        );
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
-    }
-
-    /// Calling a simple foreign call function that takes any string input, concatenates it with itself, and reverses the concatenation
-    #[test]
-    fn foreign_call_opcode_vector_input_and_output() {
-        let r_input_pointer = MemoryAddress::direct(0);
-        let r_input_size = MemoryAddress::direct(1);
-        // We need to pass a location of appropriate size
-        let r_output_pointer = MemoryAddress::direct(2);
-        let r_output_size = MemoryAddress::direct(3);
-
-        // Our first string to use the identity function with
-        let input_string: Vec<FieldElement> =
-            vec![(1u128).into(), (2u128).into(), (3u128).into(), (4u128).into()];
-        // Double the string (concatenate it with itself)
-        let mut output_string: Vec<_> =
-            input_string.iter().cloned().chain(input_string.clone()).collect();
-        // Reverse the concatenated string
-        output_string.reverse();
-
-        // First call:
-        let string_double_program = vec![
-            Opcode::Const {
-                destination: MemoryAddress::direct(100),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(input_string.len() as u32),
-            },
-            Opcode::Const {
-                destination: MemoryAddress::direct(101),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(0u64),
-            },
-            Opcode::CalldataCopy {
-                destination_address: MemoryAddress::direct(4),
-                size_address: MemoryAddress::direct(100),
-                offset_address: MemoryAddress::direct(101),
-            },
-            // input_pointer = 4
-            Opcode::Const {
-                destination: r_input_pointer,
-                value: (4u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // input_size = input_string.len() (constant here)
-            Opcode::Const {
-                destination: r_input_size,
-                value: input_string.len().into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output_pointer = 4 + input_size
-            Opcode::Const {
-                destination: r_output_pointer,
-                value: (4 + input_string.len()).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output_size = input_size * 2
-            Opcode::Const {
-                destination: r_output_size,
-                value: (input_string.len() * 2).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output_pointer[0..output_size] = string_double(input_pointer[0...input_size])
-            Opcode::ForeignCall {
-                function: "string_double".into(),
-                destinations: vec![ValueOrArray::HeapVector(HeapVector {
-                    pointer: r_output_pointer,
-                    size: r_output_size,
-                })],
-                destination_value_types: vec![HeapValueType::Vector {
-                    value_types: vec![HeapValueType::field()],
-                }],
-                inputs: vec![ValueOrArray::HeapVector(HeapVector {
-                    pointer: r_input_pointer,
-                    size: r_input_size,
-                })],
-                input_value_types: vec![HeapValueType::Vector {
-                    value_types: vec![HeapValueType::field()],
-                }],
-            },
-        ];
-
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm =
-            brillig_execute_and_get_vm(input_string.clone(), &string_double_program, &solver);
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "string_double".into(),
-                inputs: vec![input_string.clone().into()]
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(ForeignCallResult {
-            values: vec![ForeignCallParam::Array(output_string.clone())],
-        });
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check result in memory
-        let result_values: Vec<_> = vm
-            .memory
-            .read_slice(MemoryAddress::direct(4 + input_string.len()), output_string.len())
-            .iter()
-            .map(|mem_val| mem_val.clone().to_field())
-            .collect();
-        assert_eq!(result_values, output_string);
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
-    }
-
-    #[test]
-    fn foreign_call_opcode_memory_alloc_result() {
-        let r_input = MemoryAddress::direct(0);
-        let r_output = MemoryAddress::direct(1);
-
-        // Define a simple 2x2 matrix in memory
-        let initial_matrix: Vec<FieldElement> =
-            vec![(1u128).into(), (2u128).into(), (3u128).into(), (4u128).into()];
-
-        // Transpose of the matrix (but arbitrary for this test, the 'correct value')
-        let expected_result: Vec<FieldElement> =
-            vec![(1u128).into(), (3u128).into(), (2u128).into(), (4u128).into()];
-
-        let invert_program = vec![
-            Opcode::Const {
-                destination: MemoryAddress::direct(100),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(initial_matrix.len() as u32),
-            },
-            Opcode::Const {
-                destination: MemoryAddress::direct(101),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(0u64),
-            },
-            Opcode::CalldataCopy {
-                destination_address: MemoryAddress::direct(2),
-                size_address: MemoryAddress::direct(100),
-                offset_address: MemoryAddress::direct(101),
-            },
-            // input = 0
-            Opcode::Const {
-                destination: r_input,
-                value: (2u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output = 0
-            Opcode::Const {
-                destination: r_output,
-                value: (6u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // *output = matrix_2x2_transpose(*input)
-            Opcode::ForeignCall {
-                function: "matrix_2x2_transpose".into(),
-                destinations: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_output,
-                    size: initial_matrix.len(),
-                })],
-                destination_value_types: vec![HeapValueType::Array {
-                    size: initial_matrix.len(),
-                    value_types: vec![HeapValueType::field()],
-                }],
-                inputs: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_input,
-                    size: initial_matrix.len(),
-                })],
-                input_value_types: vec![HeapValueType::Array {
-                    size: initial_matrix.len(),
-                    value_types: vec![HeapValueType::field()],
-                }],
-            },
-        ];
-
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm = brillig_execute_and_get_vm(initial_matrix.clone(), &invert_program, &solver);
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "matrix_2x2_transpose".into(),
-                inputs: vec![initial_matrix.clone().into()]
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(expected_result.clone().into());
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check initial memory still in place
-        let initial_values: Vec<_> = vm
-            .memory
-            .read_slice(MemoryAddress::direct(2), 4)
-            .iter()
-            .map(|mem_val| mem_val.clone().to_field())
-            .collect();
-        assert_eq!(initial_values, initial_matrix);
-
-        // Check result in memory
-        let result_values: Vec<_> = vm
-            .memory
-            .read_slice(MemoryAddress::direct(6), 4)
-            .iter()
-            .map(|mem_val| mem_val.clone().to_field())
-            .collect();
-        assert_eq!(result_values, expected_result);
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
-    }
-
-    #[test]
-    fn foreign_call_opcode_multiple_array_inputs_result() {
-        let r_input_a = MemoryAddress::direct(0);
-        let r_input_b = MemoryAddress::direct(1);
-        let r_output = MemoryAddress::direct(2);
-
-        // Define a simple 2x2 matrix in memory
-        let matrix_a: Vec<FieldElement> =
-            vec![(1u128).into(), (2u128).into(), (3u128).into(), (4u128).into()];
-
-        let matrix_b: Vec<FieldElement> =
-            vec![(10u128).into(), (11u128).into(), (12u128).into(), (13u128).into()];
-
-        // Transpose of the matrix (but arbitrary for this test, the 'correct value')
-        let expected_result: Vec<FieldElement> =
-            vec![(34u128).into(), (37u128).into(), (78u128).into(), (85u128).into()];
-
-        let matrix_mul_program = vec![
-            Opcode::Const {
-                destination: MemoryAddress::direct(100),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(matrix_a.len() + matrix_b.len()),
-            },
-            Opcode::Const {
-                destination: MemoryAddress::direct(101),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(0u64),
-            },
-            Opcode::CalldataCopy {
-                destination_address: MemoryAddress::direct(3),
-                size_address: MemoryAddress::direct(100),
-                offset_address: MemoryAddress::direct(101),
-            },
-            // input = 3
-            Opcode::Const {
-                destination: r_input_a,
-                value: (3u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // input = 7
-            Opcode::Const {
-                destination: r_input_b,
-                value: (7u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // output = 0
-            Opcode::Const {
-                destination: r_output,
-                value: (0u128).into(),
-                bit_size: BitSize::Integer(MEMORY_ADDRESSING_BIT_SIZE),
-            },
-            // *output = matrix_2x2_transpose(*input)
-            Opcode::ForeignCall {
-                function: "matrix_2x2_transpose".into(),
-                destinations: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_output,
-                    size: matrix_a.len(),
-                })],
-                destination_value_types: vec![HeapValueType::Array {
-                    size: matrix_a.len(),
-                    value_types: vec![HeapValueType::field()],
-                }],
-                inputs: vec![
-                    ValueOrArray::HeapArray(HeapArray { pointer: r_input_a, size: matrix_a.len() }),
-                    ValueOrArray::HeapArray(HeapArray { pointer: r_input_b, size: matrix_b.len() }),
-                ],
-                input_value_types: vec![
-                    HeapValueType::Array {
-                        size: matrix_a.len(),
-                        value_types: vec![HeapValueType::field()],
-                    },
-                    HeapValueType::Array {
-                        size: matrix_b.len(),
-                        value_types: vec![HeapValueType::field()],
-                    },
-                ],
-            },
-        ];
-        let mut initial_memory = matrix_a.clone();
-        initial_memory.extend(matrix_b.clone());
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm = brillig_execute_and_get_vm(initial_memory, &matrix_mul_program, &solver);
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "matrix_2x2_transpose".into(),
-                inputs: vec![matrix_a.into(), matrix_b.into()]
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(expected_result.clone().into());
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check result in memory
-        let result_values: Vec<_> = vm
-            .memory
-            .read_slice(MemoryAddress::direct(0), 4)
-            .iter()
-            .map(|mem_val| mem_val.clone().to_field())
-            .collect();
-        assert_eq!(result_values, expected_result);
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
-    }
-
-    #[test]
-    fn foreign_call_opcode_nested_arrays_and_slices_input() {
-        // [(1, <2,3>, [4]), (5, <6,7,8>, [9])]
-
-        let v2: Vec<MemoryValue<FieldElement>> = vec![
-            MemoryValue::new_field(FieldElement::from(2u128)),
-            MemoryValue::new_field(FieldElement::from(3u128)),
-        ];
-        let a4: Vec<MemoryValue<FieldElement>> =
-            vec![MemoryValue::new_field(FieldElement::from(4u128))];
-        let v6: Vec<MemoryValue<FieldElement>> = vec![
-            MemoryValue::new_field(FieldElement::from(6u128)),
-            MemoryValue::new_field(FieldElement::from(7u128)),
-            MemoryValue::new_field(FieldElement::from(8u128)),
-        ];
-        let a9: Vec<MemoryValue<FieldElement>> =
-            vec![MemoryValue::new_field(FieldElement::from(9u128))];
-
-        // construct memory by declaring all inner arrays/vectors first
-        // Declare v2
-        let v2_ptr: usize = 0usize;
-        let mut memory = vec![MemoryValue::from(1_u32), v2.len().into()];
-        memory.extend(v2.clone());
-        let a4_ptr = memory.len();
-        memory.extend(vec![MemoryValue::from(1_u32)]);
-        memory.extend(a4.clone());
-        let v6_ptr = memory.len();
-        memory.extend(vec![MemoryValue::from(1_u32), v6.len().into()]);
-        memory.extend(v6.clone());
-        let a9_ptr = memory.len();
-        memory.extend(vec![MemoryValue::from(1_u32)]);
-        memory.extend(a9.clone());
-        // finally we add the contents of the outer array
-        memory.extend(vec![MemoryValue::from(1_u32)]);
-        let outer_start = memory.len();
-        let outer_array = vec![
-            MemoryValue::new_field(FieldElement::from(1u128)),
-            MemoryValue::from(v2.len() as u32),
-            MemoryValue::from(v2_ptr),
-            MemoryValue::from(a4_ptr),
-            MemoryValue::new_field(FieldElement::from(5u128)),
-            MemoryValue::from(v6.len() as u32),
-            MemoryValue::from(v6_ptr),
-            MemoryValue::from(a9_ptr),
-        ];
-        memory.extend(outer_array.clone());
-
-        let input_array_value_types: Vec<HeapValueType> = vec![
-            HeapValueType::field(),
-            HeapValueType::Simple(BitSize::Integer(IntegerBitSize::U64)), // size of following vector
-            HeapValueType::Vector { value_types: vec![HeapValueType::field()] },
-            HeapValueType::Array { value_types: vec![HeapValueType::field()], size: 1 },
-        ];
-
-        // memory address of the end of the above data structures
-        let r_ptr = memory.len();
-
-        let r_input = MemoryAddress::direct(r_ptr);
-        let r_output = MemoryAddress::direct(r_ptr + 1);
-
-        let program: Vec<_> = vec![
-            Opcode::Const {
-                destination: MemoryAddress::direct(100),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(memory.len()),
-            },
-            Opcode::Const {
-                destination: MemoryAddress::direct(101),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-                value: FieldElement::from(0u64),
-            },
-            Opcode::CalldataCopy {
-                destination_address: MemoryAddress::direct(0),
-                size_address: MemoryAddress::direct(100),
-                offset_address: MemoryAddress::direct(101),
-            },
-        ]
-        .into_iter()
-        .chain(memory.iter().enumerate().map(|(index, mem_value)| Opcode::Cast {
-            destination: MemoryAddress::direct(index),
-            source: MemoryAddress::direct(index),
-            bit_size: mem_value.bit_size(),
-        }))
-        .chain(vec![
-            // input = 0
-            Opcode::Const {
-                destination: r_input,
-                value: (outer_start).into(),
-                bit_size: BitSize::Integer(IntegerBitSize::U32),
-            },
-            // some_function(input)
-            Opcode::ForeignCall {
-                function: "flat_sum".into(),
-                destinations: vec![ValueOrArray::MemoryAddress(r_output)],
-                destination_value_types: vec![HeapValueType::field()],
-                inputs: vec![ValueOrArray::HeapArray(HeapArray {
-                    pointer: r_input,
-                    size: outer_array.len(),
-                })],
-                input_value_types: vec![HeapValueType::Array {
-                    value_types: input_array_value_types,
-                    size: outer_array.len(),
-                }],
-            },
-        ])
-        .collect();
-
-        let solver = StubbedBlackBoxSolver::default();
-        let mut vm = brillig_execute_and_get_vm(
-            memory.into_iter().map(|mem_value| mem_value.to_field()).collect(),
-            &program,
-            &solver,
-        );
-
-        // Check that VM is waiting
-        assert_eq!(
-            vm.status,
-            VMStatus::ForeignCallWait {
-                function: "flat_sum".into(),
-                inputs: vec![ForeignCallParam::Array(vec![
-                    (1u128).into(),
-                    (2u128).into(), // size of following vector
-                    (2u128).into(),
-                    (3u128).into(),
-                    (4u128).into(),
-                    (5u128).into(),
-                    (3u128).into(), // size of following vector
-                    (6u128).into(),
-                    (7u128).into(),
-                    (8u128).into(),
-                    (9u128).into(),
-                ])],
-            }
-        );
-
-        // Push result we're waiting for
-        vm.resolve_foreign_call(FieldElement::from(45u128).into());
-
-        // Resume VM
-        brillig_execute(&mut vm);
-
-        // Check that VM finished once resumed
-        assert_eq!(vm.status, VMStatus::Finished { return_data_offset: 0, return_data_size: 0 });
-
-        // Check result
-        let result_value = vm.memory.read(r_output);
-        assert_eq!(result_value, MemoryValue::new_field(FieldElement::from(45u128)));
-
-        // Ensure the foreign call counter has been incremented
-        assert_eq!(vm.foreign_call_counter, 1);
     }
 
     #[test]
